requirements.txt
alembic==1.18.4
amqp==5.3.1
bcrypt==5.0.0
billiard==4.2.4
blinker==1.9.0
celery==5.6.2
certifi==2026.2.25
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Deprecated==1.3.1
Flask==3.1.3
Flask-Bcrypt==1.0.1
flask-cors==6.0.2
Flask-JWT-Extended==4.7.1
Flask-Limiter==4.1.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.3.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.6.2
limits==5.8.0
Mako==1.3.10
MarkupSafe==3.0.3
marshmallow==4.2.2
ordered-set==4.1.0
packaging==26.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
PyJWT==2.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
redis==7.2.1
requests==2.32.5
six==1.17.0
SQLAlchemy==2.0.47
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
Werkzeug==3.1.6
wrapt==2.1.1
